Move GeneralSettings into its own container component
