+++ app/javascript/react/screens/App/Settings/index.js
@@ -1,24 +1,16 @@
 import { connect } from 'react-redux';
 import Settings from './Settings';
-
-import * as SettingsActions from './SettingsActions';
-import * as NotificationActions from '../common/NotificationList/NotificationListActions';
-import * as RouterActions from '../../../../redux/actions/routerActions';
-
 import reducer from './SettingsReducer';
+import * as RouterActions from '../../../../redux/actions/routerActions';
 
 export const reducers = { settings: reducer };
 
-const mapStateToProps = ({ settings, form }, ownProps) => ({
-  ...settings,
-  ...ownProps.data,
-  settingsForm: form.settings
-});
+const mapStateToProps = () => ({});
 
 const mergeProps = (stateProps, dispatchProps, ownProps) => Object.assign(stateProps, ownProps.data, dispatchProps);
 
 export default connect(
   mapStateToProps,
-  Object.assign(SettingsActions, NotificationActions, RouterActions),
+  RouterActions,
   mergeProps
 )(Settings);
