+++ app/javascript/react/screens/App/Settings/screens/ConversionHostsSettings/index.js
@@ -1,0 +1,3 @@
+import ConversionHostsSettings from './ConversionHostsSettings';
+
+export default ConversionHostsSettings;
